+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Future breaking changes
 
 - Change minimal version to PHP 7.2
+- Parameter to negate DBAL::sqlIsNull will be deprecated
 
 # Version 2.2.0 DEVELOPMENT
 
 - Fix parameter name on DBAL::sqlFieldEscape from $tableName to $fieldName
+- Add DBAL::sqlIsNotNull as a negative to DBAL::sqlIsNull
+- Add notice if using DBAL::sqlIsNull with negation
 - Move test from driver specification to DbalCommonSqlTrait on testSqlIsNull & testSqlIfNull
 
 # Version 2.1.3 2018-10-05